# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion des dettes et créances",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Gestion des dettes et récues"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Lister ensemble des articles du stock",
                "description": "Liste tous les articles avec filtre optionnel de disponibilité (rôle Boutiquier requis). Retourne 'Liste des articles' si des articles sont trouvés, sinon 'Pas Articles' avec data=null",
                "operationId": "89181b74a8aa4d580750ede6aa32f92a",
                "parameters": [
                    {
                        "name": "disponible",
                        "in": "query",
                        "description": "Filtrer par disponibilité: oui (qteStock > 0) ou non (qteStock = 0)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "oui",
                                "non"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès - Liste des articles ou 'Pas Articles' si vide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Article"
                                                    }
                                                },
                                                {
                                                    "type": "null"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "description": "'Liste des articles' si trouvés, 'Pas Articles' si vide",
                                            "type": "string",
                                            "example": "Liste des articles"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Enregistrer un nouvel article",
                "description": "Crée un nouvel article (rôle Boutiquier requis)",
                "operationId": "ce06e2544b80bc76b184c771c0fb2e7e",
                "requestBody": {
                    "description": "Données de l'article à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Article"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article enregistrer avec succees"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "411": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 411
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/articles/libelle": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Lister un article a partir de son libelle",
                "operationId": "791dc57b67e356a4086dcb24c55d1530",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Lait Laicran 700g"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article trouve",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Article"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "article trouve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "411": {
                        "description": "Objet non trouve",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 411
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Objet non trouve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/articles/restore/{id}": {
            "patch": {
                "tags": [
                    "Articles"
                ],
                "summary": "Restore a soft-deleted article",
                "operationId": "04a1e1313e26cbdbaf494f365e7a1802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article to restore",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article restored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Article"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreArticleRequest": {
                "required": [
                    "libelle",
                    "prix",
                    "qteStock"
                ],
                "properties": {
                    "libelle": {
                        "description": "Libellé de l'article (obligatoire et unique)",
                        "type": "string",
                        "example": "Lait Laicran 700g"
                    },
                    "prix": {
                        "description": "Prix de l'article (obligatoire)",
                        "type": "number",
                        "format": "float",
                        "example": 1700
                    },
                    "qteStock": {
                        "description": "Quantité en stock (obligatoire)",
                        "type": "integer",
                        "example": 1000
                    }
                },
                "type": "object"
            },
...
            "Article": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "libelle": {
                        "type": "string",
                        "example": "Article Example"
                    },
                    "prix": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "quantite": {
                        "type": "integer",
                        "format": "int32",
                        "example": 100
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}